Unit test fix
git-svn-id: https://svn.apache.org/repos/asf/db/ddlutils/trunk@685049 13f79535-47bb-0310-9956-ffa450edef68

+++ src/test/java/org/apache/ddlutils/io/TestInsertColumn.java
@@ -1876,7 +1876,7 @@ public class TestInsertColumn extends TestAgainstLiveDatabaseBase
         List beans2 = getRows("roundtrip2");
 
         assertEquals(new Integer(1), beans1.get(0), "pk");
-        assertEquals(new Integer(1), beans2.get(0), "pk");
+        assertEquals(new Integer(2), beans2.get(0), "pk");
         assertEquals(new Integer(1), beans2.get(0), "avalue");
     }
